Manque View Html/css

+++ P2-BDE-Events/Models/Prestations/Prestation.cs
@@ -7,7 +7,7 @@ namespace P2_BDE_Events.Models.Prestations
         public int Id { get; set; }
         public string Titre { get; set; }
         public int CapaciteMax { get; set; }
-        public decimal Tarif { get; set; }
+        public double Tarif { get; set; }
         public string Calendrier { get; set; }
         public bool Livraison { get; set; }
         public string Description { get; set; }
